allow author download, add permissionservice, fixing tests
