+++ test/endpoint/councilmeeting.spec.js
@@ -93,8 +93,8 @@ test('councilmeeting get', async (t) => {
         `responseMeetingArray is ${typeof responseMeetingArray}, maybe its length is 0`)
 
     const responseMeeting = responseMeetingArray.find(meeting =>
-        Object.keys(councilMeeting).find(key =>
-            councilMeeting[key] === meeting[key])
+        Object.keys(councilMeeting).every(key =>
+            councilMeeting[key].toString() === meeting[key].toString())
     )
     t.truthy(responseMeeting,
         `councilMeeting was not found in responseMeetingArray, ${responseMeeting}`)
